requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-jose
passlib[bcrypt]
requests